Subclassed tableview cell and made a dashboard cell, defined table view delegate and datasource methods in dashboard viewcontroller
