--- Bluebird Heli/DashboardViewController.swift
@@ -1,46 +1,0 @@
-//
-//  DashboardViewController.swift
-//  Bluebird Heli
-//
-//  Created by Sean Calkins on 11/21/17.
-//  Copyright © 2017 Sean Calkins. All rights reserved.
-//
-
-import UIKit
-
-class DashboardViewController: UIViewController {
-
-    @IBOutlet var tableView: UITableView!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-}
-
-extension DashboardViewController: UITableViewDataSource {
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        return UITableViewCell()
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 0
-    }
-    
-}
-
-extension DashboardViewController: UITableViewDelegate {
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        print("hi")
-    }
-    
-}
