--- Bluebird Heli/DateCell.swift
@@ -1,16 +1,0 @@
-//
-//  DateCell.swift
-//  Bluebird Heli
-//
-//  Created by Sean Calkins on 11/21/17.
-//  Copyright © 2017 Sean Calkins. All rights reserved.
-//
-
-import UIKit
-import JTAppleCalendar
-
-class DateCell: JTAppleCell {
-    
-    @IBOutlet var dateLabel: UILabel!
-    
-}
