--- Bluebird Heli/GoogleService-Info.plist
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>AD_UNIT_ID_FOR_BANNER_TEST</key>
-	<string>ca-app-pub-3940256099942544/2934735716</string>
-	<key>AD_UNIT_ID_FOR_INTERSTITIAL_TEST</key>
-	<string>ca-app-pub-3940256099942544/4411468910</string>
-	<key>CLIENT_ID</key>
-	<string>434807084951-5cpbal10vduth8n3qnneojqsb0g3o6qk.apps.googleusercontent.com</string>
-	<key>REVERSED_CLIENT_ID</key>
-	<string>com.googleusercontent.apps.434807084951-5cpbal10vduth8n3qnneojqsb0g3o6qk</string>
-	<key>API_KEY</key>
-	<string>AIzaSyA59ZumhSOtECgaGVjXI4SYF38HnV1Fw2Y</string>
-	<key>GCM_SENDER_ID</key>
-	<string>434807084951</string>
-	<key>PLIST_VERSION</key>
-	<string>1</string>
-	<key>BUNDLE_ID</key>
-	<string>com.cloudVeil.bluebirdHeli</string>
-	<key>PROJECT_ID</key>
-	<string>bluebirdheli-dd1f5</string>
-	<key>STORAGE_BUCKET</key>
-	<string>bluebirdheli-dd1f5.appspot.com</string>
-	<key>IS_ADS_ENABLED</key>
-	<true></true>
-	<key>IS_ANALYTICS_ENABLED</key>
-	<false></false>
-	<key>IS_APPINVITE_ENABLED</key>
-	<false></false>
-	<key>IS_GCM_ENABLED</key>
-	<true></true>
-	<key>IS_SIGNIN_ENABLED</key>
-	<true></true>
-	<key>GOOGLE_APP_ID</key>
-	<string>1:434807084951:ios:bf2bc589b04ec793</string>
-	<key>DATABASE_URL</key>
-	<string>https://bluebirdheli-dd1f5.firebaseio.com</string>
-</dict>
-</plist>
