--- Bluebird Heli/LoginViewController.swift
@@ -1,36 +1,0 @@
-//
-//  LoginViewController.swift
-//  Bluebird Heli
-//
-//  Created by Sean Calkins on 11/21/17.
-//  Copyright © 2017 Sean Calkins. All rights reserved.
-//
-
-import UIKit
-
-class LoginViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-    @IBAction func loginTapped(_ sender: UIButton) {
-       self.showDashboard()
-    }
-    
-    func showDashboard() {
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        let rootViewController = storyboard.instantiateViewController(withIdentifier: "Dashboard") as! UINavigationController
-        if let appDelegate = UIApplication.shared.delegate as? AppDelegate {
-            appDelegate.window?.rootViewController = rootViewController
-        }
-    }
-
-}
