+++ Bluebird Heli/Scheduling/View Controllers/SchedulingViewController.swift
@@ -1,0 +1,55 @@
+//
+//  SchedulingViewController.swift
+//  Bluebird Heli
+//
+//  Created by Sean Calkins on 11/21/17.
+//  Copyright © 2017 Sean Calkins. All rights reserved.
+//
+
+import UIKit
+import JTAppleCalendar
+
+class SchedulingViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+
+}
+
+extension SchedulingViewController: JTAppleCalendarViewDataSource {
+    
+    func configureCalendar(_ calendar: JTAppleCalendarView) -> ConfigurationParameters {
+       
+        let startDate = Date().addingTimeInterval(-31536000)
+        let endDate = Date().addingTimeInterval(31536000)
+        
+        let parameters = ConfigurationParameters(startDate: startDate, endDate: endDate)
+        
+        return parameters
+        
+    }
+    
+}
+
+extension SchedulingViewController: JTAppleCalendarViewDelegate {
+    
+    func calendar(_ calendar: JTAppleCalendarView, willDisplay cell: JTAppleCell, forItemAt date: Date, cellState: CellState, indexPath: IndexPath) {
+    }
+    
+    func calendar(_ calendar: JTAppleCalendarView, cellForItemAt date: Date, cellState: CellState, indexPath: IndexPath) -> JTAppleCell {
+        let cell = calendar.dequeueReusableJTAppleCell(withReuseIdentifier: "dateCell", for: indexPath) as! DateCell
+        cell.dateLabel.text = cellState.text
+        
+        return cell
+    }
+
+    
+}
